% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat.R
\name{hisat_build_usage}
\alias{hisat_build_usage}
\title{Print available arguments for hisat_build_usage}
\usage{
hisat_build_usage()
}
\value{
hisat_build available arguments and their usage.
}
\description{
Note that some arguments to the
hisat_build_usage are invalid if they are
already handled as explicit function arguments.
}
\examples{
hisat_build_usage()
}
\references{
Kim, D., Langmead, B. & Salzberg, S. HISAT: a fast 
spliced aligner with low memory requirements. Nat Methods 12, 
357-360 (2015).
}
\author{
Qixiu Du
}
