% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat.R
\name{hisat_inspect}
\alias{hisat_inspect}
\title{Interface to hisat-inspect of hisat-0.1.6-beta}
\usage{
hisat_inspect(bt2Index, ...)
}
\arguments{
\item{...}{Additional arguments to be passed on to the binaries.
See below for details.}

\item{bt2Index}{\code{Character} scalar.name of any of the index files 
but with the .X.bt2 or .rev.X.bt2 suffix omitted.}
}
\value{
a FASTA file containing the sequences of the original references 
(with all non-A/C/G/T characters converted to Ns)
}
\description{
This function can be use to call wrapped \code{hisat-inspect}
binary.
}
\examples{
td <- tempdir()
## Building a hisat index
refs <- dir(system.file(package="Rhisat", "extdata", "bt2","refs"),
full=TRUE)
hisat_build(references=refs, bt2Index=file.path(td, "lambda_virus"),
"--quiet",overwrite=TRUE)
if(file.exists(file.path(td, "lambda_virus.1.bt2"))){
    cmdout<-hisat_inspect(bt2Index = file.path(td, "lambda_virus"),
       "--summary");cmdout 
}

}
\references{
Langmead, B., & Salzberg, S. L. (2012).
Fast gapped-read alignment with Bowtie 2. Nature methods, 9(4), 357-359.
}
\author{
QiXiu Du
}
