% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat.R
\name{hisat_usage}
\alias{hisat_usage}
\title{Print available arguments for hisat}
\usage{
hisat_usage()
}
\value{
hisat available arguments and their usage.
}
\description{
Note that some arguments to the
hisat are invalid if they are
already handled as explicit function arguments.
}
\examples{
hisat_usage()
}
\references{
Langmead, B., & Salzberg, S. L. (2012). Fast gapped-read
alignment with Bowtie 2. Nature methods, 9(4), 357-359.
}
\author{
Qixiu Du
}
