% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat.R
\name{hisat_inspect}
\alias{hisat_inspect}
\title{Interface to hisat-inspect of hisat-0.1.6-beta}
\usage{
hisat_inspect(bt2Index, ..., exe = TRUE)
}
\arguments{
\item{bt2Index}{\code{Character} scalar.name of any of the index files 
but with the .X.bt2 or .rev.X.bt2 suffix omitted.}

\item{...}{Additional arguments to be passed on to the binaries.
See below for details.}

\item{exe}{\code{Logical}. Define whether execute the command if setting
\code{TRUE}. Else it will return the command without execution.}
}
\value{
a FASTA file containing the sequences of the original references 
(with all non-A/C/G/T characters converted to Ns)
}
\description{
This function can be use to call wrapped \code{hisat-inspect}
binary.
}
\examples{
td <- tempdir()
td <- gsub("[\\\\]","/",td)
## Building a hisat index
refs <- dir(system.file(package="Rhisat", "extdata", "bt2","refs"),
full=TRUE)
hisat_build(references=refs, bt2Index=file.path(td, "lambda_virus"),
"--quiet",overwrite=TRUE)
Sys.sleep(2)
if(file.exists(file.path(td, "lambda_virus.1.bt2"))){
    cmdout<-hisat_inspect(bt2Index = file.path(td, "lambda_virus"),
       "--summary");cmdout 
}

}
\references{
Kim, D., Langmead, B. & Salzberg, S. HISAT: a fast 
spliced aligner with low memory requirements. Nat Methods 12, 
357-360 (2015).
}
\author{
QiXiu Du
}
