% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat.R
\name{hisat_inspect_usage}
\alias{hisat_inspect_usage}
\title{Print available arguments for hisat_inspect_usage}
\usage{
hisat_inspect_usage()
}
\value{
hisat_inspect available arguments and their usage.
}
\description{
Note that some arguments to the
hisat_inspect_usage are invalid if they are
already handled as explicit function arguments.
}
\examples{
hisat_inspect_usage()
}
\references{
Langmead B, Salzberg S.
Fast gapped-read alignment with Bowtie 2. Nature Methods. 2012, 9:357-359.
}
\author{
Qixiu Du
}
