% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat.R
\name{hisat_version}
\alias{hisat_version}
\title{Print version information of hisat-0.1.6-beta}
\usage{
hisat_version()
}
\value{
An invisible \code{Integer} of call status.
The value is 0 when there is not any mistakes
Otherwise the value is non-zero.
}
\description{
Print version information of hisat-0.1.6-beta
}
\examples{
cmdout<-hisat_version();cmdout
}
\references{
Langmead, B., & Salzberg, S. L. (2012). Fast gapped-read
alignment with Bowtie 2. Nature methods, 9(4), 357-359.
}
\author{
Qixiu Du
}
